Create GC32 STUDENT CREATED: THE PRISONERS DILEMMA .py
The prisoner's dilemma is when the two prisoners are given the option to either confess or stay silent. If one of them stays silent and the other one confesses, the confessor is released from prison and the silent one is in prison for 20 years. If they both stay silent they are both in prison for only 1 year each. If they both confess then they are both in prison for 5 years each. Design a program that simulates this with the user being prisoner A and the computer being prisoner B. It must have the computer choose different outcomes and the player should have the choice to play again.