Create GC33 Diet Calculator .py
The program should ask the user for input of the name of the food he/she has consumed and the amount of calories in the food then, it displays the amount of calories the user needs to consume to stay healthy for the day. The program should have a filter for invalid inputs and it should show the stats and reset when the user types "end day".