Create AS32 Improved Security .py
It has been discovered that burglars are taking advantage of the fact that security guards start at the top of the building and wind their way down. To improve security in the building, guards now need to go to randomly assigned number. However every floor needs to be visited. Create a random generator that will make sure guards visit every floor once in a random order. Building managers should be able to choose the number of floors their building has.